# bank-backend/api-calls/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/transfers/{id}/cancel": {
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "cancelTransfer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/verify-account": {
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "verifyAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyAccountRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VerifyAccountResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/scheduled": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getScheduledTransfers",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransferResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "scheduledTransfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduledTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/recurring": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getRecurringTransfers",
        "parameters": [
          {
            "name": "activeOnly",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageRecurringTransferResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "recurringTransfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringTransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/recurring/{id}/cancel": {
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "cancelRecurringTransfer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/internal": {
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "internalTransfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InternalTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/external": {
      "post": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "externalTransfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transfers": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getAllTransfers",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/{id}": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getTransfer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/{id}/receipt": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getReceipt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferReceiptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/statistics": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getStatistics",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransferStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transfers/pending": {
      "get": {
        "tags": [
          "transfer-controller"
        ],
        "operationId": "getPendingTransfers",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageTransferResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/transactions/{id}/receipt": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getReceipt_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionReceiptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/transactions/statistics": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "operationId": "getStatistics_1",
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/{id}/receipt": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getReceipt_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentReceiptResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/payments/statistics": {
      "get": {
        "tags": [
          "payment-controller"
        ],
        "operationId": "getStatistics_2",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/cards/statistics": {
      "get": {
        "tags": [
          "card-controller"
        ],
        "operationId": "getStatistics_3",
        "parameters": [
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/CardStatisticsResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "VerifyAccountRequest": {
        "type": "object",
        "properties": {
          "accountNumber": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": [
          "accountNumber"
        ]
      },
      "VerifyAccountResponse": {
        "type": "object",
        "properties": {
          "isValid": {
            "type": "boolean"
          },
          "accountNumber": {
            "type": "string"
          },
          "accountType": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ScheduledTransferRequest": {
        "type": "object",
        "properties": {
          "fromAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "toAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "toAccountNumber": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "scheduledDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "amount",
          "fromAccountId",
          "scheduledDate"
        ]
      },
      "TransferResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "fromAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "toAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "transferType": {
            "type": "string"
          },
          "amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "reference": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "scheduledDate": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "RecurringTransferRequest": {
        "type": "object",
        "properties": {
          "fromAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "toAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "toAccountNumber": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "frequency": {
            "type": "string",
            "minLength": 1,
            "pattern": "DAILY|WEEKLY|BIWEEKLY|MONTHLY|QUARTERLY|YEARLY"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "amount",
          "frequency",
          "fromAccountId",
          "startDate"
        ]
      },
...
      "InternalTransferRequest": {
        "type": "object",
        "properties": {
          "fromAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "toAccountId": {
            "type": "integer",
            "format": "int64"
          },
          "amount": {
            "type": "number",
            "minimum": 0.01
          },
          "description": {
            "type": "string",
            "maxLength": 200,
            "minLength": 0
          }
        },
        "required": [
          "amount",
          "fromAccountId",
          "toAccountId"
        ]
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
...
      "PageTransferResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TransferResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "PageRecurringTransferResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecurringTransferResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}